Reasonable output
Made a new program that pretty much just cuts at "comprising" and uses everything after split by semicolons. We try to take out non-relevant details like wherein. SpaCy is used to get the NP and the remaining clause is treated as requirments, that is probably what needs to be altered... more specificity.